requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg==3.1.17
psycopg-binary==3.1.17
python-dotenv==1.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
